requirements.txt
selenium==4.16.0
webdriver-manager==4.0.1
python-dotenv==1.0.0
pyyaml==6.0.1
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
gspread==5.12.4
requests==2.31.0
beautifulsoup4==4.12.2
lxml==6.0.2
pandas
python-telegram-bot==20.7
undetected-chromedriver==3.5.5
cryptography==41.0.7
Flask==2.3.3
openai==1.3.0
pdfplumber==0.10.3
python-docx==0.8.11